More cleanup of PWC
